Create fragment app to separate the logic for queries
