+++ api/fragment/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+from fragment.models import Fragment
+
+# Register your models here.
+admin.site.register(Fragment)
