+++ api/text/admin.py
@@ -1,7 +1,5 @@
 from django.contrib import admin
-from text.models.text_model import Text
-from text.models.fragment_model import Fragment
+from text.models import Text
 
 # Register your models here.
 admin.site.register(Text)
-admin.site.register(Fragment)
